Fixed #702 and #710 Implement database and attachment encryption

+++ src/main/java/com/couchbase/lite/BlobStore.java
@@ -17,7 +17,10 @@
 
 package com.couchbase.lite;
 
+import com.couchbase.lite.support.security.SymmetricKey;
+import com.couchbase.lite.support.security.SymmetricKeyException;
 import com.couchbase.lite.util.Log;
+import com.couchbase.lite.util.TextUtils;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -27,6 +30,7 @@ import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.RandomAccessFile;
+import java.nio.file.Files;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
@@ -41,33 +45,97 @@ import java.util.zip.GZIPInputStream;
  * @exclude
  */
 public class BlobStore {
-
-    public static String FILE_EXTENSION = ".blob";
-    public static String TMP_FILE_EXTENSION = ".blobtmp";
-    public static String TMP_FILE_PREFIX = "tmp";
+    public static final String FILE_EXTENSION = ".blob";
+    public static final String TMP_FILE_EXTENSION = ".blobtmp";
+    public static final String TMP_FILE_PREFIX = "tmp";
+    public static final String ENCRYPTION_ALGORITHM = "AES";
+    public static final String ENCRYPTION_MARKER_FILENAME = "_encryption";
 
     private String path;
+    private SymmetricKey encryptionKey;
 
-    public BlobStore(String path) {
-        this(path, false);
+    public BlobStore(String path, SymmetricKey encryptionKey) throws CouchbaseLiteException {
+        this(path, encryptionKey, false);
     }
 
-    public BlobStore(String path, boolean autoMigrate) {
+    public BlobStore(String path, SymmetricKey encryptionKey, boolean autoMigrate)
+            throws CouchbaseLiteException {
         this.path = path;
-        File directory = new File(path);
+        this.encryptionKey = encryptionKey;
 
-        if (!directory.exists()) {
-            if (!directory.mkdirs())
-                Log.w(Log.TAG_DATABASE, "Unable to make directory: %s", directory);
-        }
-        if (!directory.isDirectory()) {
-            throw new IllegalStateException(String.format("Unable to create directory for: %s",
-                    directory));
+        File directory = new File(path);
+        if (directory.exists()) {
+            if (!directory.isDirectory())
+                throw new CouchbaseLiteException("BlobStore: Blobstore is not a directory", Status.ATTACHMENT_ERROR);
+            verifyExistingStore();
+        } else {
+            if (!directory.mkdirs()) {
+                Log.w(Log.TAG_DATABASE, "BlobStore: Unable to make directory: %s", directory);
+                throw new CouchbaseLiteException("Unable to create a blobstore", Status.ATTACHMENT_ERROR);
+            }
+            if (encryptionKey != null)
+                markEncrypted(true);
         }
 
         // migrate blobstore filenames.
         if (autoMigrate) {
             migrateBlobstoreFilenames(directory);
+        }
+    }
+
+    private void verifyExistingStore() throws CouchbaseLiteException {
+        File markerFile = new File(path, ENCRYPTION_MARKER_FILENAME);
+        boolean isMarkerExists = markerFile.exists();
+        String encryptionAlg = null;
+        if (isMarkerExists) {
+            try {
+                encryptionAlg = TextUtils.read(markerFile);
+            } catch (IOException e) {
+                throw new CouchbaseLiteException(e.getCause(), Status.BAD_ATTACHMENT);
+            }
+        }
+
+        if (encryptionAlg != null) {
+            // "_encryption" file is present, so make sure we support its format & have a key:
+            if (encryptionKey == null) {
+                Log.w(Log.TAG_DATABASE, "BlobStore: Opening encrypted blob-store without providing a key");
+                throw new CouchbaseLiteException(Status.UNAUTHORIZED);
+            } else if (!encryptionAlg.equals(ENCRYPTION_ALGORITHM)) {
+                Log.w(Log.TAG_DATABASE, "BlobStore: Blob store uses unrecognized encryption '" +
+                        encryptionAlg + "'");
+                throw new CouchbaseLiteException(Status.UNAUTHORIZED);
+            }
+        } else if (!isMarkerExists) {
+            // No "_encryption" file was found, so on-disk store isn't encrypted:
+            if (encryptionKey != null) {
+                // This shouldn't be happening:
+                Log.e(Log.TAG_DATABASE, "BlobStore: Blob store was corrupted. Encryption marker file was not found");
+                throw new CouchbaseLiteException(Status.CORRUPT_ERROR);
+            }
+        }
+    }
+
+    private void markEncrypted(boolean encrypted) throws CouchbaseLiteException {
+        File markerFile = new File(path, ENCRYPTION_MARKER_FILENAME);
+        if (encrypted) {
+            try {
+                TextUtils.write(ENCRYPTION_ALGORITHM, markerFile);
+                if (!markerFile.exists()) {
+                    Log.w(Log.TAG_DATABASE, "BlobStore: Unable to save the encryption marker file into the blob store");
+                }
+            } catch (IOException e) {
+                Log.w(Log.TAG_DATABASE, "BlobStore: Unable to save the encryption marker file into the blob store");
+                throw new CouchbaseLiteException(e.getCause(), Status.ATTACHMENT_ERROR);
+            }
+        } else {
+            if (markerFile.exists()) {
+                if (!markerFile.delete()) {
+                    Log.w(Log.TAG_DATABASE, "BlobStore: Unable to delete the encryption marker file in the blob store");
+                    throw new CouchbaseLiteException(
+                            "Unable to delete the encryption marker file in the Blob-store",
+                            Status.ATTACHMENT_ERROR);
+                }
+            }
         }
     }
 
@@ -95,7 +163,7 @@ public class BlobStore {
         try {
             md = MessageDigest.getInstance("SHA-1");
         } catch (NoSuchAlgorithmException e) {
-            Log.e(Log.TAG_BLOB_STORE, "Error, SHA-1 getDigest is unavailable.");
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error, SHA-1 getDigest is unavailable.");
             return null;
         }
         byte[] sha1hash = new byte[40];
@@ -110,7 +178,7 @@ public class BlobStore {
         try {
             md = MessageDigest.getInstance("SHA-1");
         } catch (NoSuchAlgorithmException e) {
-            Log.e(Log.TAG_BLOB_STORE, "Error, SHA-1 getDigest is unavailable.");
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error, SHA-1 getDigest is unavailable.");
             return null;
         }
         byte[] sha1hash = new byte[40];
@@ -125,7 +193,7 @@ public class BlobStore {
             }
             fis.close();
         } catch (IOException e) {
-            Log.e(Log.TAG_BLOB_STORE, "Error readin tmp file to compute key");
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error reading tmp file to compute key");
         }
 
         sha1hash = md.digest();
@@ -133,7 +201,14 @@ public class BlobStore {
         return result;
     }
 
+    /**
+     * Path to file storing the blob. Returns null if the blob is encrypted.
+     * @param key
+     * @return path to file storing the blob.
+     */
     public String getBlobPathForKey(BlobKey key) {
+        if (encryptionKey != null)
+            return null;
         return getRawPathForKey(key);
     }
 
@@ -154,8 +229,8 @@ public class BlobStore {
             File lowercaseFile = new File(lowercaseFilename);
             if (lowercaseFile.exists()) {
                 filename = lowercaseFilename;
-                Log.w(Log.TAG_BLOB_STORE,
-                        "Found the older attachment blobstore file. Recommend to set auto migration:\n" +
+                Log.w(Log.TAG_DATABASE,
+                        "BlobStore: Found the older attachment blobstore file. Recommend to set auto migration:\n" +
                                 "\tManagerOptions options = new ManagerOptions();\n" +
                                 "\toptions.setAutoMigrateBlobStoreFilename(true);\n" +
                                 "\tManager manager = new Manager(..., options);\n"
@@ -192,15 +267,29 @@ public class BlobStore {
     }
 
     public byte[] blobForKey(BlobKey key) {
+        if (key == null)
+            return null;
+
         String path = getRawPathForKey(key);
         File file = new File(path);
-        byte[] result = null;
+        byte[] blob = null;
         try {
-            result = getBytesFromFile(file);
+            blob = getBytesFromFile(file);
+            if (encryptionKey != null && blob != null)
+                blob = encryptionKey.decryptData(blob);
+            BlobKey decodedKey = BlobStore.keyForBlob(blob);
+            if (!key.equals(decodedKey)) {
+                Log.w(Log.TAG_DATABASE, "BlobStore: Attachment " + path + " decoded incorrectly!");
+                blob = null;
+            }
         } catch (IOException e) {
-            Log.e(Log.TAG_BLOB_STORE, "Error reading file", e);
+            blob = null;
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error reading file", e);
+        } catch (SymmetricKeyException e) {
+            blob = null;
+            Log.e(Log.TAG_DATABASE, "BlobStore: Attachment " + path + " decoded incorrectly!", e);
         }
-        return result;
+        return blob;
     }
 
     public InputStream blobStreamForKey(BlobKey key) {
@@ -208,17 +297,25 @@ public class BlobStore {
         File file = new File(path);
         if (file.canRead()) {
             try {
-                return new FileInputStream(file);
+                InputStream is = new FileInputStream(file);
+                if (encryptionKey != null)
+                    return encryptionKey.decryptStream(is);
+                else
+                    return is;
             } catch (FileNotFoundException e) {
-                Log.e(Log.TAG_BLOB_STORE, "Unexpected file not found in blob store", e);
+                Log.e(Log.TAG_DATABASE, "BlobStore: Unexpected file not found in blob store", e);
+                return null;
+            } catch (SymmetricKeyException e) {
+                Log.e(Log.TAG_DATABASE, "BlobStore: Attachment stream " + path + " cannot be decoded!", e);
                 return null;
             }
         }
         return null;
     }
 
+    /*
+    NO Usages: Should be removed:
     public boolean storeBlobStream(InputStream inputStream, BlobKey outKey) {
-
         File tmp = null;
         try {
             tmp = File.createTempFile(TMP_FILE_PREFIX, TMP_FILE_EXTENSION, new File(path));
@@ -232,7 +329,7 @@ public class BlobStore {
             inputStream.close();
             fos.close();
         } catch (IOException e) {
-            Log.e(Log.TAG_BLOB_STORE, "Error writing blog to tmp file", e);
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error writing blob to tmp file", e);
             return false;
         }
 
@@ -251,6 +348,7 @@ public class BlobStore {
         }
         return true;
     }
+    */
 
     public boolean storeBlob(byte[] data, BlobKey outKey) {
         BlobKey newKey = keyForBlob(data);
@@ -261,15 +359,28 @@ public class BlobStore {
             return true;
         }
 
+        if (encryptionKey != null) {
+            SymmetricKeyException error = null;
+            try {
+                data = encryptionKey.encryptData(data);
+            } catch (SymmetricKeyException e) {
+                error = e;
+            }
+            if (data == null) {
+                Log.w(Log.TAG_DATABASE, "BlobStore: Failed to encode data for " + path, error);
+                return false;
+            }
+        }
+
         FileOutputStream fos = null;
         try {
             fos = new FileOutputStream(file);
             fos.write(data);
         } catch (FileNotFoundException e) {
-            Log.e(Log.TAG_BLOB_STORE, "Error opening file for output", e);
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error opening file for output", e);
             return false;
         } catch (IOException ioe) {
-            Log.e(Log.TAG_BLOB_STORE, "Error writing to file", ioe);
+            Log.e(Log.TAG_DATABASE, "BlobStore: Error writing to file", ioe);
             return false;
         } finally {
             if (fos != null) {
@@ -347,13 +458,14 @@ public class BlobStore {
         File[] contents = file.listFiles();
         for (File attachment : contents) {
             BlobKey attachmentKey = new BlobKey();
-            getKeyForFilename(attachmentKey, attachment.getPath());
-            if (!keysToKeep.contains(attachmentKey)) {
-                boolean result = attachment.delete();
-                if (result) {
-                    ++numDeleted;
-                } else {
-                    Log.e(Log.TAG_BLOB_STORE, "Error deleting attachment: %s", attachment);
+            if (getKeyForFilename(attachmentKey, attachment.getPath())) {
+                if (!keysToKeep.contains(attachmentKey)) {
+                    boolean result = attachment.delete();
+                    if (result) {
+                        ++numDeleted;
+                    } else {
+                        Log.e(Log.TAG_BLOB_STORE, "Error deleting attachment: %s", attachment);
+                    }
                 }
             }
         }
@@ -392,5 +504,13 @@ public class BlobStore {
         }
 
         return tempDirectory;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public SymmetricKey getEncryptionKey() {
+        return encryptionKey;
     }
 }
